refactor(search): split searcher.py into a package
The 2200-line hybrid search module is now mempalace/searcher/ so PRs
can target ranking, sqlite BM25, the pipeline, or CLI output without
colliding.

The public import path is unchanged: fragments exec into the package
namespace so search, search_memories, SearchError, and test
monkeypatches on process globals keep working.
